requirements.txt
setuptools
numpy
pandas
python-docx
docopt